Cleaned up the entire project
