--- src/client/components/index.tsx
@@ -1,2 +1,0 @@
-export { LoginForm } from "./LoginForm";
-export { LogoutFab } from "./LogoutFab";
