--- src/client/styles/Clients.module.css
@@ -1,4 +1,0 @@
-.list {
-	width: 100%;
-	max-width: 480px;
-}
