+++ src/client/types/CarePlan.tsx
@@ -1,0 +1,10 @@
+import { Goal } from "./Goal";
+
+interface CarePlan {
+	_id: string;
+	Id: string;
+	DisplayName: string;
+	ClientId: string;
+	Goals: Array<Goal>;
+}
+export type { CarePlan };
