+++ src/client/types/Client.tsx
@@ -1,0 +1,9 @@
+interface Client {
+	_id: string;
+	FirstName: string;
+	LastName: string;
+	BirthDate: string;
+	Gender: string;
+	NativeId: string;
+}
+export type { Client };
