+++ src/client/types/ClientListProps.tsx
@@ -1,0 +1,7 @@
+import { Client } from "./Client";
+
+interface ClientListProps {
+	employeeClients: Array<Client>;
+	onChange: (id: string) => void;
+}
+export type { ClientListProps };
