+++ src/client/types/Employee.tsx
@@ -1,0 +1,7 @@
+interface Employee {
+	_id: string;
+	Name: string;
+	EmployeeId: string;
+	AuthorizedClients: Array<string>;
+}
+export type { Employee };
