+++ src/client/types/Goal.tsx
@@ -1,0 +1,5 @@
+interface Goal {
+	DisplayName: string;
+	GoalId: string;
+}
+export type { Goal };
