+++ src/client/types/LoginData.tsx
@@ -1,0 +1,5 @@
+interface LoginData {
+	id: string;
+	message: string;
+}
+export type { LoginData };
