+++ src/client/types/index.ts
@@ -1,0 +1,7 @@
+export * from "./CarePlan";
+export * from "./Client";
+export * from "./ClientListProps";
+export * from "./Employee";
+export * from "./Goal";
+export * from "./LoginData";
+export * from "./Report";
